+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TestTableRepairJob.java
@@ -22,21 +22,30 @@ import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.ignoreStubs;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static org.mockito.Mockito.when;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
+import com.datastax.driver.core.Host;
 import com.ericsson.bss.cassandra.ecchronos.core.JmxProxyFactory;
 import com.ericsson.bss.cassandra.ecchronos.core.MockedClock;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.state.RepairState;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.state.RepairStateSnapshot;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.state.ReplicaRepairGroup;
+import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReference;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -166,11 +175,11 @@ public class TestTableRepairJob
     {
         // mock
         doReturn(true).when(myRepairStateSnapshot).canRepair();
-
+        mockRepairGroup(0L);
         assertThat(myRepairJob.runnable()).isTrue();
 
         verify(myRepairState, times(1)).update();
-        verify(myRepairStateSnapshot, times(1)).canRepair();
+        verify(myRepairStateSnapshot, times(2)).canRepair();
     }
 
     @Test
@@ -178,12 +187,12 @@ public class TestTableRepairJob
     {
         // mock
         doReturn(false).doReturn(true).when(myRepairStateSnapshot).canRepair();
-
+        mockRepairGroup(0L);
         assertThat(myRepairJob.runnable()).isFalse();
         assertThat(myRepairJob.runnable()).isTrue();
 
         verify(myRepairState, times(2)).update();
-        verify(myRepairStateSnapshot, times(2)).canRepair();
+        verify(myRepairStateSnapshot, times(3)).canRepair();
     }
 
     @Test
@@ -294,6 +303,7 @@ public class TestTableRepairJob
         // mock - not repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         assertThat(myRepairJob.runnable()).isTrue();
@@ -308,6 +318,7 @@ public class TestTableRepairJob
         // mock - repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         myRepairJob.runnable();
@@ -337,6 +348,7 @@ public class TestTableRepairJob
         // mock - not repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         assertThat(myRepairJob.runnable()).isTrue();
@@ -351,6 +363,7 @@ public class TestTableRepairJob
         // mock - repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         assertThat(myRepairJob.runnable()).isFalse();
@@ -374,6 +387,7 @@ public class TestTableRepairJob
         // mock - not repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         assertThat(myRepairJob.runnable()).isTrue();
@@ -388,6 +402,7 @@ public class TestTableRepairJob
         // mock - repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         myRepairJob.postExecute(true);
@@ -411,6 +426,7 @@ public class TestTableRepairJob
         // mock - not repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         assertThat(myRepairJob.runnable()).isTrue();
@@ -425,6 +441,7 @@ public class TestTableRepairJob
         // mock - repaired
         doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
         myClock.setTime(start);
 
         myRepairJob.runnable();
@@ -467,6 +484,7 @@ public class TestTableRepairJob
         // We have waited 2 days to repair, send alarm and run repair
         doReturn(lastRepairedAtWarning).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepairedAtWarning);
 
         assertThat(myRepairJob.runnable()).isTrue();
 
@@ -478,6 +496,7 @@ public class TestTableRepairJob
         // Repair has been completed
         doReturn(lastRepairedAtAfterRepair).when(myRepairStateSnapshot).lastRepairedAt();
         doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepairedAtAfterRepair);
 
         myRepairJob.postExecute(true);
 
@@ -494,7 +513,6 @@ public class TestTableRepairJob
         verify(myFaultReporter, times(0)).raise(any(FaultCode.class), anyMapOf(String.class, Object.class));
     }
 
-
     @Test
     public void testGetView()
     {
@@ -503,5 +521,71 @@ public class TestTableRepairJob
         assertThat(repairJobView.getTableReference()).isEqualTo(myTableReference);
         assertThat(repairJobView.getRepairConfiguration()).isEqualTo(myRepairConfiguration);
         assertThat(repairJobView.getRepairStateSnapshot()).isEqualTo(myRepairStateSnapshot);
+    }
+
+    @Test
+    public void testGetRealPriority()
+    {
+        long lastRepaired = System.currentTimeMillis();
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
+        doReturn(false).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(-1);
+
+        lastRepaired = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) - TimeUnit.HOURS.toMillis(1));
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(-1);
+
+        lastRepaired = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS));
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(1);
+
+        lastRepaired = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) + TimeUnit.HOURS.toMillis(1));
+        doReturn(lastRepaired).when(myRepairStateSnapshot).lastRepairedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        mockRepairGroup(lastRepaired);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(2);
+    }
+
+    @Test
+    public void testGetRealPrioritySnapshotLastRepairedAtLowerThanRepairGroups()
+    {
+        long lastRepairedAtSnapshot = System.currentTimeMillis() - TimeUnit.DAYS.toMillis(14);
+        doReturn(lastRepairedAtSnapshot).when(myRepairStateSnapshot).lastRepairedAt();
+        doReturn(true).when(myRepairStateSnapshot).canRepair();
+        long firstRepairGroupLastRepairedAt = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS));
+        ReplicaRepairGroup firstReplicaRepairGroup = getRepairGroup(new LongTokenRange(1, 2), firstRepairGroupLastRepairedAt);
+        mockRepairGroup(firstReplicaRepairGroup);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(1);
+
+        long secondRepairGroupLastRepairedAt = System.currentTimeMillis() - (TimeUnit.DAYS.toMillis(RUN_INTERVAL_IN_DAYS) + TimeUnit.HOURS.toMillis(1));
+        ReplicaRepairGroup secondReplicaRepairGroup = getRepairGroup(new LongTokenRange(2, 3), secondRepairGroupLastRepairedAt);
+        mockRepairGroup(secondReplicaRepairGroup, firstReplicaRepairGroup);
+        assertThat(myRepairJob.getRealPriority()).isEqualTo(2);
+    }
+
+    private void mockRepairGroup(long lastRepairedAt)
+    {
+        mockRepairGroup(getRepairGroup(new LongTokenRange(1, 2), lastRepairedAt));
+    }
+
+    private void mockRepairGroup(ReplicaRepairGroup ...replicaRepairGroups)
+    {
+        List<ReplicaRepairGroup> repairGroups = new ArrayList<>();
+        for (ReplicaRepairGroup replicaRepairGroup : replicaRepairGroups)
+        {
+            repairGroups.add(replicaRepairGroup);
+        }
+        when(myRepairStateSnapshot.getRepairGroups()).thenReturn(repairGroups);
+    }
+
+    private ReplicaRepairGroup getRepairGroup(LongTokenRange range, long lastRepairedAt)
+    {
+        ImmutableSet<Host> replicas = ImmutableSet.of(mock(Host.class), mock(Host.class));
+        return new ReplicaRepairGroup(replicas, ImmutableList.of(range), lastRepairedAt);
     }
 }